backend/scraper/requirements.txt
requests
python-dateutil
pytz
peewee==3.15
#psycopg2==2.9
